Add try/rescue for BadMapError
FeederEx does not support proper return values, but will instead just
raise an exception whenever it can't parse, meaning there is no good way
to catch the error.

In the future I might check for the erroneous data myself, but I'm
skeptical, because it'll just be another form of error handling anyway,
and checking data before hand is brittle.